requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.6.0
    # via django
django==4.1.4
    # via
    #   -r requirements.in
    #   django-fbv
    #   django-markdownx
    #   django-markdowny
    #   django-tagulous
django-fbv==0.5.1
    # via -r requirements.in
django-markdownx==4.0.0b1
    # via -r requirements.in
django-markdowny==0.2.4
    # via -r requirements.in
django-tagulous==1.3.3
    # via -r requirements.in
markdown==3.4.1
    # via
    #   django-markdownx
    #   django-markdowny
mysqlclient==2.1.1
    # via -r requirements.in
pillow==9.4.0
    # via django-markdownx
sqlparse==0.4.3
    # via django
